requirements.txt
annotated-types==0.7.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
